requirements.txt
langgraph
langgraph-checkpoint-sqlite
requests
pypdf
ebooklib
beautifulsoup4
watchdog
python-dotenv
